# backend/internal/api/server/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: KubeCoder API
  version: 0.1.0
  description: KubeCoder control-plane API.
servers:
  - url: /
...
paths:
...
  /api/v1/organizations:
    get:
      tags: [Organizations]
      operationId: listUserOrganizations
      summary: List organizations accessible to the authenticated user
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Query' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Accessible organizations
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /api/v1/organizations/{slug}:
    parameters:
      - { $ref: '#/components/parameters/OrganizationSlug' }
    get:
      tags: [Organizations]
      operationId: getUserOrganization
      summary: Return an organization accessible to the authenticated user
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Accessible organization
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/organizations/{slug}/preferred:
    parameters:
      - { $ref: '#/components/parameters/OrganizationSlug' }
    put:
      tags: [Organizations]
      operationId: setPreferredOrganization
      summary: Set the authenticated user preferred organization
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Preferred organization
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Organization' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/organizations/{slug}/ssh-key:
    parameters:
      - { $ref: '#/components/parameters/OrganizationSlug' }
    get:
      tags: [Organizations]
      operationId: getOrganizationSSHKey
      summary: Return the organization public SSH key
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Organization public SSH key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationSSHKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/organizations/{slug}/ssh-key/regenerate:
    parameters:
      - { $ref: '#/components/parameters/OrganizationSlug' }
    post:
      tags: [Organizations]
      operationId: regenerateOrganizationSSHKey
      summary: Create a missing organization SSH key pair or replace the existing pair
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Generated organization public SSH key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationSSHKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/v1/admin/agents:
    get:
      tags: [Agent administration]
      operationId: listAgents
      summary: List full agent definitions
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Query' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - { name: active, in: query, schema: { type: boolean } }
      responses:
        '200': { description: Agent page, content: { application/json: { schema: { $ref: '#/components/schemas/AgentPage' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Agent administration]
      operationId: createAgent
      summary: Create an agent definition
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AgentInput' } } } }
      responses:
        '201': { description: Created agent, content: { application/json: { schema: { $ref: '#/components/schemas/Agent' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/admin/agents/{id}:
    parameters: [{ $ref: '#/components/parameters/ResourceID' }]
    get:
      tags: [Agent administration]
      operationId: getAgent
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Agent definition, content: { application/json: { schema: { $ref: '#/components/schemas/Agent' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Agent administration]
      operationId: updateAgent
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AgentInput' } } } }
      responses:
        '200': { description: Updated agent, content: { application/json: { schema: { $ref: '#/components/schemas/Agent' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Agent administration]
      operationId: deleteAgent
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Agent deleted }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/organizations/{slug}/agents:
    parameters: [{ $ref: '#/components/parameters/OrganizationSlug' }]
    get:
      tags: [Repositories]
      operationId: listOrganizationAgents
      summary: List active public agents available to an organization member
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Query' }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200': { description: Public agent page, content: { application/json: { schema: { $ref: '#/components/schemas/PublicAgentPage' } } } }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  parameters:
    Query:
      name: query
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
...
    OrganizationSlug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
    ResourceID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrganizationSummary:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
    Organization:
      allOf:
        - { $ref: '#/components/schemas/OrganizationSummary' }
        - type: object
          required: [created_at, updated_at]
          properties:
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...
    OrganizationPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Organization' }
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
    OrganizationSSHKey:
      type: object
      description: Public metadata for an organization's encrypted SSH identity. Private key material is never returned.
      required: [public_key, fingerprint, algorithm, created_at, updated_at]
      properties:
        public_key: { type: string, pattern: '^ssh-ed25519 ' }
        fingerprint: { type: string, pattern: '^SHA256:' }
        algorithm: { type: string, enum: [ssh-ed25519] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    PublicAgent:
      type: object
      required: [id, name, description, capabilities, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        capabilities: { type: array, items: { type: string } }
        active: { type: boolean }
    AgentInput:
      type: object
      required: [name, description, runtime_adapter, runtime_version, image, provider, model, system_prompt, capabilities, cpu_millis, memory_mb, storage_mb, max_duration_seconds, active]
      properties:
        name: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 1000 }
        runtime_adapter: { type: string }
        runtime_version: { type: string }
        image: { type: string }
        provider: { type: string }
        model: { type: string }
        system_prompt: { type: string }
        capabilities: { type: array, items: { type: string } }
        cpu_millis: { type: integer, minimum: 1 }
        memory_mb: { type: integer, minimum: 1 }
        storage_mb: { type: integer, minimum: 1 }
        max_duration_seconds: { type: integer, minimum: 1 }
        active: { type: boolean }
      additionalProperties: false
    Agent:
      allOf:
        - { $ref: '#/components/schemas/AgentInput' }
        - type: object
          required: [id, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    AgentPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Agent' } }
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
    PublicAgentPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/PublicAgent' } }
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
...
    Problem:
      type: object
      required: [type, title, status, detail, instance, correlation_id]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        correlation_id: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Access forbidden
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: Resource conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...